separated BucketTreeItem into Raster and Vector
